@@ -1,274 +1,186 @@
 """
-OrderedDict is part _collections Python > 3.5
-We need pure python version of it to implement insertion.
+Utililities for pyparam
 """
-# pylint: skip-file
-import _collections_abc
-from _weakref import proxy as _proxy
-from reprlib import recursive_repr as _recursive_repr
+import re
+import textwrap
 
-class _Link:
-	__slots__ = 'prev', 'next', 'key', '__weakref__'
+class _Valuable:
 
-class _OrderedDictKeysView(_collections_abc.KeysView):
+    STR_METHODS = ('capitalize',
+                   'center',
+                   'count',
+                   'decode',
+                   'encode',
+                   'endswith',
+                   'expandtabs',
+                   'find',
+                   'format',
+                   'index',
+                   'isalnum',
+                   'isalpha',
+                   'isdigit',
+                   'islower',
+                   'isspace',
+                   'istitle',
+                   'isupper',
+                   'join',
+                   'ljust',
+                   'lower',
+                   'lstrip',
+                   'partition',
+                   'replace',
+                   'rfind',
+                   'rindex',
+                   'rjust',
+                   'rpartition',
+                   'rsplit',
+                   'rstrip',
+                   'split',
+                   'splitlines',
+                   'startswith',
+                   'strip',
+                   'swapcase',
+                   'title',
+                   'translate',
+                   'upper',
+                   'zfill')
 
-	def __reversed__(self):
-		yield from reversed(self._mapping)
+    def __str__(self):
+        return str(self.value)
 
-class _OrderedDictItemsView(_collections_abc.ItemsView):
+    def str(self):
+        """Return the value in str type"""
+        return str(self.value)
 
-	def __reversed__(self):
-		for key in reversed(self._mapping):
-			yield (key, self._mapping[key])
+    def int(self, raise_exc=True):
+        """Return the value in int type"""
+        try:
+            return int(self.value)
+        except (ValueError, TypeError):
+            if raise_exc:
+                raise
+            return None
 
-class _OrderedDictValuesView(_collections_abc.ValuesView):
+    def float(self, raise_exc=True):
+        """Return the value in float type"""
+        try:
+            return float(self.value)
+        except (ValueError, TypeError):
+            if raise_exc:
+                raise
+            return None
 
-	def __reversed__(self):
-		for key in reversed(self._mapping):
-			yield self._mapping[key]
+    def bool(self):
+        """Return the value in bool type"""
+        return bool(self.value)
 
-class OrderedDict(dict):
-	'Dictionary that remembers insertion order'
-	# An inherited dict maps keys to values.
-	# The inherited dict provides __getitem__, __len__, __contains__, and get.
-	# The remaining methods are order-aware.
-	# Big-O running times for all methods are the same as regular dictionaries.
+    def __getattr__(self, item):
+        # attach str methods
+        if item in _Valuable.STR_METHODS:
+            return getattr(str(self.value), item)
+        raise AttributeError(
+            'Class %r: No such attribute: %r' % (self.__class__.__name__, item)
+        )
 
-	# The internal self.__map dict maps keys to links in a doubly linked list.
-	# The circular doubly linked list starts and ends with a sentinel element.
-	# The sentinel element never gets deleted (this simplifies the algorithm).
-	# The sentinel is in self.__hardroot with a weakref proxy in self.__root.
-	# The prev links are weakref proxies (to prevent circular references).
-	# Individual links are kept alive by the hard reference in self.__map.
-	# Those hard references disappear when a key is deleted from an OrderedDict.
+    def __add__(self, other):
+        return self.value + other
 
-	def __init__(*args, **kwds):
-		'''Initialize an ordered dictionary.  The signature is the same as
-		regular dictionaries.  Keyword argument order is preserved.
-		'''
-		if not args:
-			raise TypeError("descriptor '__init__' of 'OrderedDict' object "
-							"needs an argument")
-		self, *args = args
-		if len(args) > 1:
-			raise TypeError('expected at most 1 arguments, got %d' % len(args))
-		try:
-			self.__root
-		except AttributeError:
-			self.__hardroot = _Link()
-			self.__root = root = _proxy(self.__hardroot)
-			root.prev = root.next = root
-			self.__map = {}
-		self.__update(*args, **kwds)
+    def __contains__(self, other):
+        return other in self.value
 
-	def __setitem__(self, key, value,
-					dict_setitem=dict.__setitem__, proxy=_proxy, Link=_Link):
-		'od.__setitem__(i, y) <==> od[i]=y'
-		# Setting a new item creates a new link at the end of the linked list,
-		# and the inherited dictionary is updated with the new key/value pair.
-		if key not in self:
-			self.__map[key] = link = Link()
-			root = self.__root
-			last = root.prev
-			link.prev, link.next, link.key = last, root, key
-			last.next = link
-			root.prev = proxy(link)
-		dict_setitem(self, key, value)
+    def __hash__(self): # pragma: no cover
+        """
+        Use id as identifier for hash
+        """
+        return id(self)
 
-	def __delitem__(self, key, dict_delitem=dict.__delitem__):
-		'od.__delitem__(y) <==> del od[y]'
-		# Deleting an existing item uses self.__map to find the link which gets
-		# removed by updating the links in the predecessor and successor nodes.
-		dict_delitem(self, key)
-		link = self.__map.pop(key)
-		link_prev = link.prev
-		link_next = link.next
-		link_prev.next = link_next
-		link_next.prev = link_prev
-		link.prev = None
-		link.next = None
+    def __eq__(self, other): # pragma: no cover
+        return self.value == other
 
-	def __iter__(self):
-		'od.__iter__() <==> iter(od)'
-		# Traverse the linked list in order.
-		root = self.__root
-		curr = root.next
-		while curr is not root:
-			yield curr.key
-			curr = curr.next
+    def __ne__(self, other):
+        return not self.__eq__(other)
 
-	def __reversed__(self):
-		'od.__reversed__() <==> reversed(od)'
-		# Traverse the linked list in reverse order.
-		root = self.__root
-		curr = root.prev
-		while curr is not root:
-			yield curr.key
-			curr = curr.prev
+class _Hashable:
+    """
+    A class for object that can be hashable
+    """
+    def __hash__(self):
+        """
+        Use id as identifier for hash
+        """
+        return id(self)
 
-	def clear(self):
-		'od.clear() -> None.  Remove all items from od.'
-		root = self.__root
-		root.prev = root.next = root
-		self.__map.clear()
-		dict.clear(self)
+    def __eq__(self, other):
+        """
+        How to compare the hash keys
+        """
+        return id(self) == id(other)
 
-	def popitem(self, last=True):
-		'''Remove and return a (key, value) pair from the dictionary.
+    def __ne__(self, other):
+        """
+        Compare hash keys
+        """
+        return not self.__eq__(other)
 
-		Pairs are returned in LIFO order if last is true or FIFO order if false.
-		'''
-		if not self:
-			raise KeyError('dictionary is empty')
-		root = self.__root
-		if last:
-			link = root.prev
-			link_prev = link.prev
-			link_prev.next = root
-			root.prev = link_prev
-		else:
-			link = root.next
-			link_next = link.next
-			root.next = link_next
-			link_next.prev = root
-		key = link.key
-		del self.__map[key]
-		value = dict.pop(self, key)
-		return key, value
+def wraptext(text,
+             width=70,
+             defaults=("Default: ", "DEFAULT:"),
+             break_long_words=False,
+             **kwargs):
+    """Wrap a text
+    # keep the indentation
+    # '  - hello world' =>
+    # '  - hello \'
+    # '    world'
+    # '  1. hello world' =>
+    # '  1. hello \'
+    # '     world'
+    """
+    width -= 2 # for ending ' \'
+    match = re.match(r'\s*(?:[-*#]|\w{1,2}\.)?\s+', text)
+    prefix = ' ' * len(match.group(0)) if match else ''
+    kwargs['subsequent_indent'] = prefix + \
+                                  kwargs.get('initial_indent', '') + \
+                                  kwargs.get('subsequent_indent', '')
 
-	def move_to_end(self, key, last=True):
-		'''Move an existing element to the end (or beginning if last is false).
+    codes = re.findall(r'`[^`]+`', text)
+    placeholders = {}
+    for i, line in enumerate(codes):
+        text = text.replace(line, '__code_%d__' % i)
+        placeholders['__code_%d__' % i] = line
 
-		Raise KeyError if the element does not exist.
-		'''
-		link = self.__map[key]
-		link_prev = link.prev
-		link_next = link.next
-		soft_link = link_next.prev
-		link_prev.next = link_next
-		link_next.prev = link_prev
-		root = self.__root
-		if last:
-			last = root.prev
-			link.prev = last
-			link.next = root
-			root.prev = soft_link
-			last.next = link
-		else:
-			first = root.next
-			link.prev = root
-			link.next = first
-			first.prev = soft_link
-			root.next = link
+    if text.endswith(' \\'):
+        return (kwargs.get('initial_indent', '') + text).splitlines()
 
-	def __sizeof__(self):
-		sizeof = _sys.getsizeof
-		n = len(self) + 1                       # number of links including root
-		size = sizeof(self.__dict__)            # instance dictionary
-		size += sizeof(self.__map) * 2          # internal dict and inherited dict
-		size += sizeof(self.__hardroot) * n     # link objects
-		size += sizeof(self.__root) * n         # proxy objects
-		return size
+    codes = textwrap.wrap(text,
+                          width,
+                          break_long_words=break_long_words,
+                          **kwargs)
 
-	update = __update = _collections_abc.MutableMapping.update
+    if not codes:
+        return codes
 
-	def keys(self):
-		"D.keys() -> a set-like object providing a view on D's keys"
-		return _OrderedDictKeysView(self)
+    default_index = None
+    for i, line in enumerate(codes):
+        for placeholder, origin in placeholders.items():
+            line = line.replace(placeholder, origin)
+        codes[i] = line
+        if defaults and any(default in line for default in defaults):
+            default_index = i
 
-	def items(self):
-		"D.items() -> a set-like object providing a view on D's items"
-		return _OrderedDictItemsView(self)
+    # put default in a separate line
+    if (default_index is not None and
+            default_index < len(codes) - 1 and
+            defaults and
+            not any(codes[default_index].startswith(default)
+                    for default in defaults)):
+        line_default_index = [codes[default_index].rfind(default)
+                              for default in defaults
+                              if default in codes[default_index]][0]
+        codes.insert(default_index+1, codes[default_index][line_default_index:])
+        codes[default_index] = codes[default_index][:line_default_index]
 
-	def values(self):
-		"D.values() -> an object providing a view on D's values"
-		return _OrderedDictValuesView(self)
-
-	__ne__ = _collections_abc.MutableMapping.__ne__
-
-	__marker = object()
-
-	def pop(self, key, default=__marker):
-		'''od.pop(k[,d]) -> v, remove specified key and return the corresponding
-		value.  If key is not found, d is returned if given, otherwise KeyError
-		is raised.
-
-		'''
-		if key in self:
-			result = self[key]
-			del self[key]
-			return result
-		if default is self.__marker:
-			raise KeyError(key)
-		return default
-
-	def setdefault(self, key, default=None):
-		'''Insert key with a value of default if key is not in the dictionary.
-
-		Return the value for key if key is in the dictionary, else default.
-		'''
-		if key in self:
-			return self[key]
-		self[key] = default
-		return default
-
-	@_recursive_repr()
-	def __repr__(self):
-		'od.__repr__() <==> repr(od)'
-		if not self:
-			return '%s()' % (self.__class__.__name__,)
-		return '%s(%r)' % (self.__class__.__name__, list(self.items()))
-
-	def __reduce__(self):
-		'Return state information for pickling'
-		inst_dict = vars(self).copy()
-		for k in vars(OrderedDict()):
-			inst_dict.pop(k, None)
-		return self.__class__, (), inst_dict or None, None, iter(self.items())
-
-	def copy(self):
-		'od.copy() -> a shallow copy of od'
-		return self.__class__(self)
-
-	@classmethod
-	def fromkeys(cls, iterable, value=None):
-		'''Create a new ordered dictionary with keys from iterable and values set to value.
-		'''
-		self = cls()
-		for key in iterable:
-			self[key] = value
-		return self
-
-	def __eq__(self, other):
-		'''od.__eq__(y) <==> od==y.  Comparison to another OD is order-sensitive
-		while comparison to a regular mapping is order-insensitive.
-
-		'''
-		if isinstance(other, OrderedDict):
-			return dict.__eq__(self, other) and list(self.keys()) == list(other.keys())
-		return dict.__eq__(self, other)
-
-	def __insertion(self, link_prev, key_value):
-		key, value = key_value
-		if link_prev == self.__root:
-			first = link_prev.next
-			self.__map[key] = first.prev = self.__root.next = _Link()
-			self.__map[key].prev = self.__root
-			self.__map[key].next = first
-			self.__map[key].key = key
-		elif link_prev.key != key:
-			if key in self:
-				del self[key]
-			link_next = link_prev.next
-			self.__map[key] = link_prev.next = link_next.prev = _Link()
-			self.__map[key].prev = link_prev
-			self.__map[key].next = link_next
-			self.__map[key].key = key
-
-		dict.__setitem__(self, key, value)
-
-	def insert_after(self, existing_key, key_value):
-		self.__insertion(self.__map[existing_key], key_value)
-
-	def insert_before(self, existing_key, key_value):
-		self.__insertion(self.__map[existing_key].prev, key_value)
+    return [(line + ' \\')
+            if i not in (len(codes)-1, default_index)
+            else line
+            for i, line in enumerate(codes)]
